requirements.txt
fastapi
uvicorn
mysql-connector-python
pandas
python-dotenv
langchain-openai
langchain-community
chromadb
pdfplumber